+++ R/add_pathlength.R
@@ -26,17 +26,26 @@ add_pathlength <- function(x) {
 #' @name add_pathlength
 #' @export
 add_pathlength.data.frame <- function(x) {
-
-  x <- hy(x)
-
-  x <- add_pathlength(x)
-
-  hy_reverse(x)
+  hy_as_dataframe(x, "add_pathlength")
 }
 
 #' @name add_pathlength
 #' @export
 add_pathlength.hy <- function(x) {
+  hy_classify_and_redispatch(x, "add_pathlength", "hy_topo", hy_guidance_topo)
+}
+
+# TODO: support hy_node auto-convert via add_toids()
+#' @name add_pathlength
+#' @export
+add_pathlength.hy_node <- function(x) {
+  hy_dispatch_error("add_pathlength", "hy_topo", x,
+    "Use add_toids() to convert fromnode/tonode to edge list.")
+}
+
+#' @name add_pathlength
+#' @export
+add_pathlength.hy_topo <- function(x) {
 
   check_names(x, c(id, toid, length_km), "add_pathlength")
 
@@ -62,5 +71,7 @@ add_pathlength.hy <- function(x) {
 
   x$pathlength_km <- pathlength_km
 
-  left_join(orig_order, x, by = id)
+  x <- left_join(orig_order, x, by = id)
+
+  classify_hy(x)
 }