+++ R/add_toids.R
@@ -54,6 +54,21 @@ add_toids.data.frame <- function(x, return_dendritic = TRUE) {
 #' @name add_toids
 #' @export
 add_toids.hy <- function(x, return_dendritic = TRUE) {
+  hy_classify_and_redispatch(x, "add_toids", "hy_node", hy_guidance_node,
+    return_dendritic = return_dendritic)
+}
+
+#' @name add_toids
+#' @export
+add_toids.hy_topo <- function(x, return_dendritic = TRUE) {
+  stop("This network already has toid (class: ", hy_network_type(x),
+    "). add_toids() converts fromnode/tonode to toid.",
+    call. = FALSE)
+}
+
+#' @name add_toids
+#' @export
+add_toids.hy_node <- function(x, return_dendritic = TRUE) {
 
   if ("toid" %in% names(x)) stop("network already contains a toid attribute")